+++ BusBookingSystem/src/main/java/com/booking/system/model/dto/TicketDTO.java
@@ -15,7 +15,7 @@ public class TicketDTO {
     private String busArrivalCity;
     private LocalDateTime busDepartureTime;
     private LocalDateTime busArrivalTime;
-    private long seat;
+    private int seat;
     private long price;
     private String userFirstName;
     private String userLastName;
